Extracted out handle change callback for use in BaseLetterButton
