+++ frontend/src/features/game/letter/button/LetterButton.tsx
@@ -1,15 +1,10 @@
 import { useState } from 'react';
-import { Button, Stack, useTheme } from '@mui/material';
-import { useIndexContext } from '../IndexProvider.tsx';
-import { useAppSelector } from '../../../../common/hooks.ts';
+import { Stack } from '@mui/material';
 import { IncrementLetterButton } from './IncrementLetterButton.tsx';
 import { DecrementLetterButton } from './DecrementLetterButton.tsx';
+import { BaseLetterButton } from './BaseLetterButton.tsx';
 
 export function LetterButton() {
-    const { index } = useIndexContext();
-    const letter = useAppSelector((state) => state.game.letters[index]);
-    const theme = useTheme();
-
     const [hovering, setHovering] = useState(false);
 
     return (
@@ -21,20 +16,7 @@ export function LetterButton() {
             onMouseLeave={() => setHovering(false)}
         >
             <IncrementLetterButton hovering={hovering} />
-            <Button
-                variant={'outlined'}
-                sx={{
-                    minWidth: 0,
-                    width: '1.25rem',
-                    height: '2.5rem',
-                    '&:disabled': {
-                        background: theme.palette.success.main,
-                        color: theme.palette.success.contrastText,
-                    },
-                }}
-            >
-                {letter}
-            </Button>
+            <BaseLetterButton />
             <DecrementLetterButton hovering={hovering} />
         </Stack>
     );
